+++ app/di.php
@@ -6,7 +6,7 @@ $registrar->addInstance(new \GovukComponents\Blocks\Button());
 $registrar->addInstance(new \GovukComponents\Blocks\Details\Block());
 $registrar->addInstance(new \GovukComponents\Blocks\InsetText\Block());
 $registrar->addInstance(new \GovukComponents\Blocks\NotificationBanner());
-$registrar->addInstance(new \GovukComponents\Blocks\WarningText());
+$registrar->addInstance(new \GovukComponents\Blocks\WarningText\Block());
 
 $registrar->addInstance(new \GovukComponents\BlockCategory());
 
@@ -18,7 +18,7 @@ $registrar->addInstance(new \GovukComponents\BlockController(
 		$registrar->getInstance(\GovukComponents\Blocks\Details\Block::class),
 		$registrar->getInstance(\GovukComponents\Blocks\InsetText\Block::class),
 		$registrar->getInstance(\GovukComponents\Blocks\NotificationBanner::class),
-		$registrar->getInstance(\GovukComponents\Blocks\WarningText::class)
+		$registrar->getInstance(\GovukComponents\Blocks\WarningText\Block::class)
 	]
 ));